Issue fixed: _ckeck function implemented, tests are working
The issue was fixed. The check code was being repeated for every
operation verification: adition, subtraction, division and
multiplication; so I took the common code out of those functions and put
it in the body of the _check function, and then just call the _check
function inside the 4 operations functions.

Then, changed the describe.only to describe.

All 22 tests are being executed, and the code passes all of them