Give better instructions
It took me a good amount of time to figure out why the tests are still marked as **Pending** even though my codes were correct.
Then i found out that there is a **.skip** on the test which mark it as pending. 

I think it will be more helpful for beginner to get a note about removing that **.skip** , otherwise they will be spending a lot of time trying to figure out why their tests are not passing.